Move test_shebang_line_detection to test_commands
